bug fixes + rearranging
